Adding event creator to list of authorized users, checking if user is authorized in event update and delete

+++ backend/src/events.py
@@ -1,7 +1,8 @@
 from backend.src.database import clear, db
 import subprocess
 import datetime
-from backend.src.error import InputError
+from backend.src.error import InputError, AccessError
+from backend.src.auth import decode_token
 from bson import ObjectId
 
 def events_crawl():
@@ -44,12 +45,15 @@ def event_is_valid(event):
     # if not is_real_date(event['deadline']): return False
     return True
 
-def event_create(event):
+def event_create(token, event):
+    user_id = decode_token(token)
     if event_already_exists(event):
         raise InputError('Event already exists')
     if not event_is_valid(event):
         raise InputError('Invalid event')
     event['ranking'] = 0
+    event['authorized_users'] = [user_id]
+    event['creator'] = user_id
     result = db.events.insert_one(event)
     return {
         'event_id': str(result.inserted_id)
@@ -59,7 +63,13 @@ def get_event(event_id):
     print(event_id)
     return db.events.find_one({ '_id': ObjectId(event_id) })
 
-def event_update(event_id, new_event):
+def user_is_authorized(user_id, event_id):
+    return user_id in db['events'].find({ '_id': ObjectId(event_id) })['authorized_users']
+
+def event_update(token, event_id, new_event):
+    user_id = decode_token(token)
+    if not user_is_authorized(user_id, event_id):
+        raise AccessError('User not authorized to update event')
     event = get_event(event_id)
     if event is None:
         raise InputError('No event in database with specified event_id')
@@ -78,9 +88,21 @@ def event_update(event_id, new_event):
     )
     return {}
 
-def event_delete(event_id):
+def user_is_creator(user_id, event_id):
+    return user_id == db['events'].find({ '_id': ObjectId(event_id) })['creator']
+
+def event_delete(token, event_id):
+    user_id = decode_token(token)
+    if not user_is_creator(user_id, event_id):
+        raise AccessError('User not authorized to delete event')
     event = get_event(event_id)
     if event is None:
         raise InputError('No event in database with specified event_id')
     db.events.delete_one({ '_id': ObjectId(event_id)})
+    return {}
+
+def event_authorize(token, event_id, to_be_added_id):
+    user_id = decode_token(token)
+    if not user_is_creator(user_id, event_id):
+        raise AccessError('User is not authorized to allow other people to manage event')
     return {}
